pyproject.toml
[tool.poetry]
package-mode = false
name = "Dungeons and Dragons"
version = "0.1.0"
description = ""
authors = ["Gal Gilor <gilorgal2010@gmail.com>"]
readme = "README.md"

[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.2"
pytest-mock = "^3.14.0"
ipykernel = "^6.29.5"

[tool.poetry.dependencies]
python = "^3.11"
dspy-ai = {extras = ["groq", "pinecone"], version = "^2.4.13"}
google-cloud-aiplatform = "^1.62.0"
google-cloud-storage = "^2.18.2"
pydantic = "^2.8.2"
jinja2 = "^3.1.4"
transformers = {extras = ["torch"], version = "^4.44.0"}
accelerate = "^0.33.0"
datasets = "^2.21.0"
fastapi = "^0.112.1"
pydantic-settings = "^2.4.0"
typing-extensions = "^4.12.2"
uvicorn = "^0.30.6"
python-multipart = "^0.0.9"
instructor = {git = "https://github.com/jxnl/instructor.git"}
vertexai = "^1.64.0"
google-generativeai = "^0.7.2"
jsonref = "^1.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[virtualenvs]
in-project = true